Refactor: getPostCount() 리팩토링 커밋 과정에서 누락된 클래스 추가

+++ src/main/java/com/ebstudy/board_v2/repository/PostDAO.java
@@ -60,9 +60,9 @@ public class PostDAO {
      * 현재 DB에 존재하는 게시글의 총 개수를 구하는 메소드
      * @return 게시글의 총 개수
      */
-    public long getPostCount() {
+    public int getPostCount() {
 
-        long count = 0L;
+        int count = 0;
 
         try {
             SqlSessionFactory sessionFactory = SessionFactory.getSessionFactory();
